Filling things out
Added more styling and html to the tagbar as well as some basic tests
for the class. Added the ability to set the tag color as well as
default tags and max tags. Added the bones of supporting sources for
tags but haven't been implemented. The example app also not shows off
a variety of the options.