Search index and starting click search items
Added additional expectations to tests to make sure that the search
index is reset whenever the search menu goes away. The code currently
violates DRY and needs to be refactored but the tests should faciliate
that in the future.

Also added a test for click on search items that currently isn't
working because I believe the blur/focus events on the input field are
eating the click/mousedown events somehow. Need to do a bit more work
there but the test is written and is currently in a failure state
until the code is fixed.