Mouse hover and click
Turns out it wasn't something weird interfering with the mouse event
it was my own stupidity, angular events are all lower case on the
dom. Tests and html have been updated for mouse clicking on search
items. Also added tests and implmentation for mousing over search
items to make sure that the active-item class is added.