Reset on add
Added code to clear the input field on add tag. This was being handled
in the html events but moved it to the ts file to be consistant on the
various ways tags can be added. Added associated tests. Note the
functional tests now include the FormsModule which may increase
testing time for them.