Obserables
Move from using an async general source to obserables for aysnc tag
data. This seems like a better plan at the moment and is working far
better in the examples. The testing still needs to be fixed though.

The default loader is unfortunately just that. Its pretty lifeless but
it does work.