Observable testing
Added testing elements to the tagbar-example project for obserable
testing.

Added a simple obserable object with a static small source.
Added an obserable object that calls out to a public api as an
example.
Many tests need to be added and there are some issues that need to be
worked out but it works largely as intended.