Keyboard events and testing improvements
Got character based keyboard events to work a bit better by using
onKeyUp event rather than the on key down event. Operational keys like
ArrowUp, Esc, etc still use keydown but character ones should get
filtered through the keyup event so that what is in the input field is
properly passed to the functions that need it. Its possible that other
keyboard events like Backspace need to go through the keyup function
as well but its a start. Updated a number of the tests to use this new
functionality and got everything back to zero. Observable testing
still remains elusive so that has been commented out for the moment
while a better testing approach can be searched.

Added a new element and styling for when there are no search results
on a filterable list. This still needs some testing I think.