List item refactoring
Holding all of the tagbar into a single component is pretty cumbersome
and needs refactoring badly. Added the list-item component that
handles the individual search items in the tag bar as they are being
searched. This may need to be expanded to adding a search group block
for added convenience but I'm not sure yet if that is going to be
adding more overhead and complexity yet. This is still a work in
progress and tests are running correctly for the outer main element
which is probably a sign of a need for better decoupling.