Bug fixing
Fixed the bulk of the testing errors that were introduced during the
list-item refactoring. There is a single functional test that no
longer passes. The actual test needs to be thought about a bit more in
terms of what the actual functionality should be so I'm going to leave
it in an error state for the moment.